2025-01-22 AC: Updates the main cron script and the list of validators.

+++ scripts/crons-dapp.bash
@@ -23,9 +23,24 @@ cd || exit 1
 curl https://api-flare-validators.flare.network/api/v1/validator -o validatorlist.json
 export jsonfile=`cat validatorlist.json`
 if jq -e . >/dev/null <<<"$jsonfile"; then
-  cp -f validatorlist.json /srv/tempo/wwwftsocan/public/
-  chown 1000:1000 /srv/tempo/wwwftsocan/public/validatorlist.json
-  rm -f validatorlist.json
+  if [[ `jq -e '.[0] | has("name")' <<<"$jsonfile"` == "true" && `jq -e '.[0] | has("nodeId")' <<<"$jsonfile"` == "true" && `jq -e '.[0] | has("lastSeen")' <<<"$jsonfile"` == "true" && `jq -e '.[0] | has("lastStatus")' <<<"$jsonfile"` == "true" && `jq -e '.[0] | has("uptime")' <<<"$jsonfile"` == "true" ]]; then
+    jsonlength=`jq length <<<"$jsonfile"`
+
+    for i in $(seq 1 $jsonlength)
+    do
+      if [[ `jq -e ".[$i] | has(\"name\")" <<<"$jsonfile"` == "false" ]]; then
+        php -r '$collection = json_decode(file_get_contents("validatorlist.json")); foreach($collection as $index => &$object) { if (!isset($object->name)) { $object->name = "Unknown"; } } file_put_contents("validatorlist.json", json_encode($collection, JSON_PRETTY_PRINT));'
+      fi
+    done
+
+    cp -f validatorlist.json /srv/tempo/wwwftsocan/public/
+    cp -f validatorlist.json /srv/tempo/wwwftsocan/data/
+    chown 1000:1000 /srv/tempo/wwwftsocan/public/validatorlist.json
+    chown 1000:1000 /srv/tempo/wwwftsocan/data/validatorlist.json
+    rm -f validatorlist.json
+  else
+    cp -f /srv/tempo/wwwftsocan/data/validatorlist.json /srv/tempo/wwwftsocan/public/
+  fi
 else
   echo "CRON - $( date ) : Failed to update validatorlist.json." >> /srv/tempo/wwwftsocan/logs/log.txt
 fi
